Move asset browser to shared-ui module
