+++ ui/projects/streampipes/platform-services/src/lib/model/assets/asset.constants.ts
