+++ ui/projects/streampipes/shared-ui/src/lib/components/asset-browser/asset-browser.component.html
@@ -47,8 +47,8 @@
                     fxFlex="100"
                 >
                     <span class="asset-creation-hint"
-                        >No assets configured yet - use assets to better
-                        organize resources!</span
+                        >No assets found - use assets to better organize
+                        resources!</span
                     >
                     <button
                         mat-button
