+++ ui/projects/streampipes/shared-ui/src/lib/components/asset-browser/asset-browser.model.ts
