+++ ui/projects/streampipes/shared-ui/src/lib/components/basic-field-description/basic-field-description.component.ts
@@ -16,7 +16,7 @@
  *
  */
 
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input } from '@angular/core';
 
 @Component({
     selector: 'sp-basic-field-description',
