+++ ui/projects/streampipes/shared-ui/src/lib/shared-ui.module.ts
@@ -45,9 +45,34 @@ import { MatSort } from '@angular/material/sort';
 import { SpExceptionDetailsComponent } from './components/sp-exception-message/exception-details/exception-details.component';
 import { SpWarningBoxComponent } from './components/warning-box/warning-box.component';
 import { SpBasicFieldDescriptionComponent } from './components/basic-field-description/basic-field-description.component';
+import { AssetBrowserToolbarComponent } from './components/asset-browser/asset-browser-toolbar/asset-browser-toolbar.component';
+import { AssetBrowserFilterComponent } from './components/asset-browser/asset-browser-toolbar/asset-browser-filter/asset-browser-filter.component';
+import { AssetBrowserComponent } from './components/asset-browser/asset-browser.component';
+import { AssetBrowserFilterLabelsComponent } from './components/asset-browser/asset-browser-toolbar/asset-browser-filter/asset-browser-filter-labels/asset-browser-filter-labels.component';
+import { AssetBrowserFilterOuterComponent } from './components/asset-browser/asset-browser-toolbar/asset-browser-filter/asset-browser-filter-outer/asset-browser-filter-outer.component';
+import { AssetBrowserFilterSitesComponent } from './components/asset-browser/asset-browser-toolbar/asset-browser-filter/asset-browser-filter-sites/asset-browser-filter-sites.component';
+import { AssetBrowserFilterTypeComponent } from './components/asset-browser/asset-browser-toolbar/asset-browser-filter/asset-browser-filter-type/asset-browser-filter-type.component';
+import { AssetBrowserHierarchyComponent } from './components/asset-browser/asset-browser-hierarchy/asset-browser-hierarchy.component';
+import { AssetBrowserNodeComponent } from './components/asset-browser/asset-browser-hierarchy/asset-browser-node/asset-browser-node.component';
+import { AssetBrowserNodeInfoComponent } from './components/asset-browser/asset-browser-hierarchy/asset-browser-node/asset-browser-node-info/asset-browser-node-info.component';
+import { MatSelectModule } from '@angular/material/select';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatMenuModule } from '@angular/material/menu';
+import { FormsModule } from '@angular/forms';
+import { MatTreeModule } from '@angular/material/tree';
 
 @NgModule({
     declarations: [
+        AssetBrowserComponent,
+        AssetBrowserFilterComponent,
+        AssetBrowserFilterLabelsComponent,
+        AssetBrowserFilterOuterComponent,
+        AssetBrowserFilterSitesComponent,
+        AssetBrowserFilterTypeComponent,
+        AssetBrowserHierarchyComponent,
+        AssetBrowserNodeComponent,
+        AssetBrowserNodeInfoComponent,
+        AssetBrowserToolbarComponent,
         ConfirmDialogComponent,
         PanelDialogComponent,
         StandardDialogComponent,
@@ -67,11 +92,16 @@ import { SpBasicFieldDescriptionComponent } from './components/basic-field-descr
     imports: [
         CommonModule,
         FlexLayoutModule,
+        FormsModule,
         MatButtonModule,
         MatDividerModule,
+        MatFormFieldModule,
         MatIconModule,
+        MatMenuModule,
+        MatSelectModule,
         MatTabsModule,
         MatTooltipModule,
+        MatTreeModule,
         PortalModule,
         OverlayModule,
         MatDialogModule,
@@ -80,6 +110,7 @@ import { SpBasicFieldDescriptionComponent } from './components/basic-field-descr
         MatSort,
     ],
     exports: [
+        AssetBrowserComponent,
         ConfirmDialogComponent,
         PanelDialogComponent,
         StandardDialogComponent,
