+++ ui/projects/streampipes/shared-ui/src/public-api.ts
@@ -26,6 +26,7 @@ export * from './lib/dialog/confirm-dialog/confirm-dialog.component';
 export * from './lib/dialog/panel-dialog/panel-dialog.component';
 export * from './lib/dialog/standard-dialog/standard-dialog.component';
 
+export * from './lib/components/asset-browser/asset-browser.component';
 export * from './lib/components/basic-header-title/header-title.component';
 export * from './lib/components/basic-inner-panel/basic-inner-panel.component';
 export * from './lib/components/basic-field-description/basic-field-description.component';
@@ -46,3 +47,4 @@ export * from './lib/services/jwt-token-storage.service';
 export * from './lib/services/current-user.service';
 export * from './lib/services/echarts-toolbox.service';
 export * from './lib/services/colorization.service';
+export * from './lib/components/asset-browser/asset-browser.service';
