+++ ui/src/app/assets/components/asset-details/asset-details-panel/asset-details-links/asset-details-links.component.ts
@@ -26,7 +26,7 @@ import {
 import { SpManageAssetLinksDialogComponent } from '../../../../dialog/manage-asset-links/manage-asset-links-dialog.component';
 import { DialogService, PanelType } from '@streampipes/shared-ui';
 import { EditAssetLinkDialogComponent } from '../../../../dialog/edit-asset-link/edit-asset-link-dialog.component';
-import { AssetConstants } from '../../../../constants/asset.constants';
+import { AssetConstants } from '../../../../../../../projects/streampipes/platform-services/src/lib/model/assets/asset.constants';
 
 @Component({
     selector: 'sp-asset-details-links',
