+++ ui/src/app/assets/dialog/asset-upload/asset-upload-dialog.component.ts
@@ -19,7 +19,7 @@
 import { Component } from '@angular/core';
 import { DialogRef } from '@streampipes/shared-ui';
 import { GenericStorageService } from '@streampipes/platform-services';
-import { AssetConstants } from '../../constants/asset.constants';
+import { AssetConstants } from '../../../../../projects/streampipes/platform-services/src/lib/model/assets/asset.constants';
 
 @Component({
     selector: 'sp-file-upload-dialog-component',
