+++ ui/src/app/configuration/sites-configuration/site-area-configuration/site-area-configuration.component.ts
@@ -22,7 +22,7 @@ import {
     GenericStorageService,
     LocationConfig,
 } from '@streampipes/platform-services';
-import { AssetConstants } from '../../../assets/constants/asset.constants';
+import { AssetConstants } from '../../../../../projects/streampipes/platform-services/src/lib/model/assets/asset.constants';
 import { MatTableDataSource } from '@angular/material/table';
 import { ManageSiteDialogComponent } from '../../dialog/manage-site/manage-site-dialog.component';
 import { DialogService, PanelType } from '@streampipes/shared-ui';
